requirements.txt
agentql
pandas
playwright
python-dotenv
asyncio
requests
spider-api>=0.1.0 